+++ actix-server/src/accept.rs
@@ -9,14 +9,11 @@ use log::{error, info};
 use mio::{Interest, Poll, Token as MioToken};
 
 use crate::server::Server;
-use crate::socket::{MioListener, SocketAddr};
+use crate::socket::MioListener;
 use crate::waker_queue::{WakerInterest, WakerQueue, WAKER_TOKEN};
 use crate::worker::{Conn, WorkerHandleAccept};
 
 struct ServerSocketInfo {
-    /// Address of socket. Mainly used for logging.
-    addr: SocketAddr,
-
     token: usize,
 
     lst: MioListener,
@@ -189,15 +186,12 @@ impl Accept {
         let sockets = socks
             .into_iter()
             .map(|(token, mut lst)| {
-                let addr = lst.local_addr();
-
                 // Start listening for incoming connections
                 poll.registry()
                     .register(&mut lst, MioToken(token), Interest::READABLE)
                     .unwrap_or_else(|e| panic!("Can not register io: {}", e));
 
                 ServerSocketInfo {
-                    addr,
                     token,
                     lst,
                     timeout: None,
@@ -370,14 +364,14 @@ impl Accept {
 
     fn register_logged(&self, info: &mut ServerSocketInfo) {
         match self.register(info) {
-            Ok(_) => info!("Resume accepting connections on {}", info.addr),
+            Ok(_) => info!("Resume accepting connections on {}", info.lst.local_addr()),
             Err(e) => error!("Can not register server socket {}", e),
         }
     }
 
     fn deregister_logged(&self, info: &mut ServerSocketInfo) {
         match self.poll.registry().deregister(&mut info.lst) {
-            Ok(_) => info!("Paused accepting connections on {}", info.addr),
+            Ok(_) => info!("Paused accepting connections on {}", info.lst.local_addr()),
             Err(e) => {
                 error!("Can not deregister server socket {}", e)
             }